Split out connection code
