+++ tests/conftest.py
@@ -6,8 +6,10 @@ import pytest
 
 @pytest.fixture(autouse=True)
 def override_time_sleep():
-    """Open a websocket connection."""
-    with patch("samsungtvws.remote.time.sleep"):
+    """Ignore time sleep in tests."""
+    with patch("samsungtvws.connection.time.sleep"), patch(
+        "samsungtvws.remote.time.sleep"
+    ):
         yield
 
 
@@ -15,6 +17,8 @@ def override_time_sleep():
 def get_connection():
     """Open a websocket connection."""
     connection = Mock()
-    with patch("samsungtvws.remote.websocket.create_connection") as connection_class:
+    with patch(
+        "samsungtvws.connection.websocket.create_connection"
+    ) as connection_class:
         connection_class.return_value = connection
         yield connection
